Zmiany języka i kosmetyczne

+++ ApplicationUpdater/Processes/SetOfflineProcess.cs
@@ -8,13 +8,14 @@ using System.Threading.Tasks;
 
 namespace ApplicationUpdater.Processes
 {
-    public class SetOnlineProcess : ProcessBase, IProcess<UpdateModel>
+    public class SetOffLineProcess : ProcessBase, IProcess<UpdateModel>
     {
-        private const string offLineFileName = "app_offline.htm";
+        private const string offLineFileName = "_app_offline.htm";
+        private const string searchPattern = "Trwa aktualizacja danych";
 
         public ProcesEventResult Process(UpdateModel model)
         {
-            if (Confirm("Czy chcesz przejść w stan ON-LINE aplikacji?") == false)
+            if (Confirm("Do you want to go OFFLINE mode?") == false)
             {
                 return null;
             }
@@ -23,10 +24,45 @@ namespace ApplicationUpdater.Processes
                 .GetFiles()
                 .SingleOrDefault(s => s.Name == $"{offLineFileName}");
 
-            File.Copy(file.FullName, Path.Combine(file.DirectoryName, $"_{offLineFileName}"));
+            if (file == null ||file.Exists == false)
+            {
+                var file2 = model.IntepubDirectory
+                .GetFiles()
+                .SingleOrDefault(s => s.Name == $"app_offline.htm");
+
+                if (file2 == null || file2.Exists == false)
+                {
+                    throw new Exception("File not found");
+                }
+                UpdateDateTimeInFile(file2);
+
+                return null;
+            }
+
+            UpdateDateTimeInFile(file);
             File.Delete(file.FullName);
 
+            UpdateProcess($"Switched application into OFFLINE mode");
+
             return null;
+
+        }
+
+        private void UpdateDateTimeInFile(FileInfo file)
+        {
+            var lines = File.ReadAllLines(file.FullName);
+
+            var index = lines.Select((s, i) => new { i, s })
+                .Where(t => t.s.Contains(searchPattern))
+                .Select(t => t.i)
+                .First();
+
+            var dateTimeValue = DateTime.Now.AddHours(1).ToString("HH:mm");
+
+            lines[index] = $"Trwa aktualizacja danych. Do godz. {dateTimeValue} NPG bedzie niedostepny. <br/><br/> W naglych przypadkach prosze o kontakt pod numerem: 695 877 795";
+
+            File.WriteAllLines(Path.Combine(file.DirectoryName, "app_offline.htm"), lines);
+           
         }
     }
 }
